FIN-28 Receivables refactoring - dynamic sticky row
