+++ src/src_V2/modules/manager/financial/constants/payable/index.js
@@ -1,0 +1,2 @@
+export * from './payable-table';
+export * from './payment-history';
