+++ src/src_V2/modules/manager/financial/constants/table-settings.js
@@ -1,9 +1,15 @@
 export const rowsPerPageOptions = [20, 50, 100];
 
-export const acceptedFileTypes = [
-  'image/png',
-  'image/jpg',
-  'image/jpeg',
-  'image/gif',
-  'application/pdf',
-];
+export const columnTypes = {
+  usualRow: 'usualRow',
+  moneyMask: 'moneyMask',
+  date: 'date',
+  custom: 'custom',
+};
+
+export const stickyFooterColumnTypes = {
+  usualRow: 'usualRow',
+  moneyMask: 'moneyMask',
+  emptyCell: 'emptyCell',
+  custom: 'custom',
+};
