+++ src/src_V2/modules/manager/financial/payables/edit/drawer-content/general-info/GeneralInfo.jsx
@@ -1,7 +1,7 @@
 import { Box, Divider, Stack } from '@mui/material';
 
 import { moneyMask } from '../../../../../../../utils';
-import { actualCountedCost, differenceCost, remainingCost } from '../../../../utils/cost';
+import { actualCountedCost, differenceCost, remainingCost } from '../../../../utils';
 import commonStyles from '../style.module.css';
 import styles from './GeneralInfo.module.css';
