+++ src/src_V2/modules/manager/financial/payables/table/custom-cells/CustomCells.jsx
@@ -3,11 +3,20 @@ import { Chip, TableCell } from '@mui/material';
 import { deepPurple, green, orange, pink } from '@mui/material/colors';
 
 import { TooltipText } from '../../../../../../components';
-import { moneyMask } from '../../../../../../utils';
+import { generateArrayByRange, moneyMask } from '../../../../../../utils';
 import { TableCellWrapper, TooltipTableCell } from '../../../components';
 import { PaymentStatus, PaymentType } from '../../../constants';
-import { colorCondition } from '../../../utils/table';
 import styles from './CustomCells.module.css';
+
+export function colorCondition(difference) {
+  if (difference > 0) {
+    return styles.positiveTableCell;
+  } else if (difference < 0) {
+    return styles.negativeTableCell;
+  } else {
+    return styles.neutralTableCell;
+  }
+}
 
 const PayableStatuses = ({ statusNum }) => {
   let color = null;
@@ -54,8 +63,12 @@ export const InvoiceAttachmentCell = ({ value }) => (
   </TableCellWrapper>
 );
 
-export const DifferenceCell = ({ value }) => (
-  <TableCell className={`${styles.tableCell} ${colorCondition(value)}`}>
+export const DifferenceCell = ({ value, styleName }) => (
+  <TableCell
+    className={`${styles.tableCell} ${colorCondition(value)} ${
+      styleName && styles[styleName]
+    }`}
+  >
     <TooltipText text={moneyMask(value)} />
   </TableCell>
 );
@@ -69,3 +82,8 @@ export const StatusCell = ({ value }) => (
 export const PaymentTypeCell = ({ value }) => (
   <TooltipTableCell rowValue={PaymentType[value]} />
 );
+
+const EmptyTableCellWrapper = ({ count }) =>
+  generateArrayByRange(1, count).map((el, index) => <TableCellWrapper key={index} />);
+
+export const EmptyCell = ({ count = 1 }) => <EmptyTableCellWrapper count={count} />;
