+++ src/src_V2/modules/manager/financial/receivables/Receivables.jsx
@@ -3,7 +3,7 @@ import { useCallback, useEffect, useState } from 'react';
 import { Layout } from '../../../../components';
 import { receivableColumns, rowsPerPageOptions } from '../constants';
 import receivablesList from '../mock/receivable';
-import { remainingCost } from '../utils/cost';
+import { remainingCost } from '../utils';
 import ControlPanel from './control-panel/ControlPanel';
 import ReceivableTable from './table/ReceivableTable';
