+++ src/src_V2/modules/manager/financial/receivables/edit/drawer-content/general-info/GeneralInfo.jsx
@@ -1,6 +1,6 @@
 import { Box, Divider, Stack } from '@mui/material';
 
-import { remainingCost } from '../../../../utils/cost';
+import { remainingCost } from '../../../../utils';
 import commonStyles from '../style.module.css';
 import styles from './GeneralInfo.module.css';
